test/python_backend/requirements.txt
# ===========================================
# FastAPI + MariaDB 연동용 Python 라이브러리
# ===========================================
# 설치: python -m pip install -r requirements.txt

# --- 웹 API ---
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# --- DB 연결 (MariaDB/MySQL) ---
# SQLAlchemy: ORM·세션 관리 (테이블 ↔ Python 객체)
sqlalchemy>=2.0.0
# PyMySQL: MariaDB/MySQL 드라이버 (SQLAlchemy가 실제 접속에 사용)
pymysql>=1.1.0
# cryptography: 암호화 연결 시 필요 (선택)
cryptography>=41.0.0

# --- 설정·유틸 ---
# .env 파일에서 DB 비밀번호 등 읽기
python-dotenv>=1.0.0

# --- ML 모델 (선택) joblib: model.pkl 사용 시 필요 ---
joblib>=1.3.0

# --- 통계적 이상 탐지 (Z-score 등) ---
scipy>=1.11.0
